+++ blacksheep/server/env.py
@@ -37,5 +37,5 @@ class EnvironmentSettings:
     def __init__(self) -> None:
         self.env = get_env()
         self.show_error_details = truthy(os.environ.get("APP_SHOW_ERROR_DETAILS", ""))
-        self.mount_auto_events = truthy(os.environ.get("APP_MOUNT_AUTO_EVENTS", ""))
+        self.mount_auto_events = truthy(os.environ.get("APP_MOUNT_AUTO_EVENTS", "1"))
         self.use_default_router = truthy(os.environ.get("APP_DEFAULT_ROUTER", "1"))