+++ itests/app_1.py
@@ -35,17 +35,17 @@ def get_static_path(file_name):
 app.serve_files(static_folder_path, discovery=True)
 
 
-@app.route("/hello-world")
+@app.router.route("/hello-world")
 async def hello_world():
     return text("Hello, World!")
 
 
-@app.route("/plain-json")
+@app.router.route("/plain-json")
 async def plain_json():
     return json({"message": "Hello, World!"})
 
 
-@app.route("/plain-json-error-simulation")
+@app.router.route("/plain-json-error-simulation")
 async def plain_json_error_simulation():
     return json({"message": "Hello, World!"}, status=500)
 
@@ -60,13 +60,13 @@ async def echo_headers(request):
     return response
 
 
-@app.route("/echo-cookies")
+@app.router.route("/echo-cookies")
 async def echo_cookies(request):
     cookies = request.cookies
     return json(cookies)
 
 
-@app.route("/set-cookie")
+@app.router.route("/set-cookie")
 async def set_cookies(name: FromQuery[str], value: FromQuery[str]):
     response = text("Setting cookie")
     response.set_cookie(Cookie(name.value, value.value))
@@ -126,7 +126,7 @@ async def echo_route_values_autobind(one, two, three):
     return json(dict(one=one, two=two, three=three))
 
 
-@app.route("/crash")
+@app.router.route("/crash")
 async def crash():
     raise CrashTest()
 
@@ -137,7 +137,7 @@ class Item:
         self.power = power
 
 
-@app.route("/echo-posted-json-autobind", methods=["POST"])
+@app.router.route("/echo-posted-json-autobind", methods=["POST"])
 async def upload_item(request, item: Item):
     assert request is not None
     assert item is not None